feat: move casl module into libs
