+++ src/api/audit-log/audit-log.controller.ts
@@ -4,7 +4,7 @@ import { Uuid } from '@/common/types/common.type';
 import { ApiAuth } from '@/decorators/http.decorators';
 import { CheckPolicies } from '@/decorators/policies.decorator';
 import { PoliciesGuard } from '@/guards/policies.guard';
-import { AppAbility } from '@/utils/ability.factory';
+import { AppAbility } from '@/libs/casl/ability.factory';
 import { AppActions, AppSubjects } from '@/utils/permissions.constant';
 import {
   Controller,
