+++ src/libs/casl/casl.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { CaslAbilityFactory } from './ability.factory';
+
+@Module({
+  providers: [CaslAbilityFactory],
+  exports: [CaslAbilityFactory],
+})
+export class CaslModule {}
