+++ src/libs/libs.module.ts
@@ -1,8 +1,9 @@
 import { Module } from '@nestjs/common';
 import { AwsModule } from './aws/aws.module';
+import { CaslModule } from './casl/casl.module';
 import { GcpModule } from './gcp/gcp.module';
 
 @Module({
-  imports: [AwsModule, GcpModule],
+  imports: [AwsModule, GcpModule, CaslModule],
 })
 export class LibsModule {}
